backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
supabase==2.9.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
